Static checks for API hash collision
