+++ applications/elf-loader/elf-lib/elf-loader.c
@@ -1,4 +1,4 @@
-#include "loader.h"
+#include "elf-loader.h"
 #include "elf.h"
 #include "../elf-loader-sys-api.h"
 #include <furi.h>
